goal-oriented methods
implemented first iteration of goal oriented methods.
Does not account for samples in the cloud
Does not account for science projects
Does not account for enemy movements and projects and does not guess
enemy goals
Accounts for discounts when turning in and taking molecules.
Does not account for discounts when taking projects